make vscode workspace compatible with ruby 3.2

+++ .devcontainer/post-create.sh
@@ -1,8 +1,14 @@
 #!/bin/sh
 cd /usr/local/redmine
-PLUGIN_NAME=redmine_logs
 
-cp plugins/${PLUGIN_NAME}/Gemfile_for_test plugins/${PLUGIN_NAME}/Gemfile 
+ln -s /workspaces/${PLUGIN_NAME} plugins/${PLUGIN_NAME}
+if [ -f plugins/${PLUGIN_NAME}/Gemfile_for_test ]
+then
+    cp plugins/${PLUGIN_NAME}/Gemfile_for_test plugins/${PLUGIN_NAME}/Gemfile 
+fi
+cp plugins/${PLUGIN_NAME}/test/fixtures/*.yml test/fixtures
+ln -s /workspaces/${PLUGIN_NAME}/.devcontainer/launch.json .vscode/launch.json
+
 bundle install 
 bundle exec rake redmine:plugins:migrate
 bundle exec rake redmine:plugins:migrate RAILS_ENV=test
